interpolation between position updates
