+++ core/src/main/java/smashdudes/ecs/components/PositionComponent.java
@@ -5,15 +5,18 @@ import smashdudes.ecs.Component;
 
 public class PositionComponent extends Component
 {
+    public final Vector2 prevPosition;
     public final Vector2 position;
 
     public PositionComponent()
     {
         position = new Vector2();
+        prevPosition = new Vector2();
     }
 
     public PositionComponent(Vector2 pos)
     {
         position = pos.cpy();
+        prevPosition = position.cpy();
     }
 }
