+++ core/src/main/java/smashdudes/ecs/systems/GameSystem.java
@@ -6,54 +6,12 @@ import smashdudes.ecs.Engine;
 import smashdudes.ecs.Entity;
 import smashdudes.ecs.events.Event;
 
-public abstract class GameSystem
+public abstract class GameSystem extends ISystem
 {
-    private final Array<Class<? extends Component>> components = new Array<>();
-    private final Array<Class<? extends Event>> events = new Array<>();
-    private boolean enabled = true;
-
-    protected final Engine engine;
-
     public GameSystem(Engine engine)
     {
-        this.engine = engine;
+        super(engine);
     }
-
-    public void registerComponentType(Class<? extends Component> clazz)
-    {
-        if(components.contains(clazz, true)) throw new IllegalStateException("System already has component registered");
-
-        components.add(clazz);
-    }
-
-    public void registerEventType(Class<? extends Event> clazz)
-    {
-        if(events.contains(clazz, true)) throw new IllegalStateException("System already has event registered");
-
-        events.add(clazz);
-    }
-
-    public final void update(float dt)
-    {
-        preUpdate();
-
-        for(Entity e : engine.getEntities(components))
-        {
-            updateEntity(e, dt);
-        }
-
-        postUpdate();
-    }
-
-    public final void receiveEvent(Event event)
-    {
-        if(events.contains(event.getClass(), true))
-        {
-            handleEvent(event);
-        }
-    }
-
-    protected void handleEvent(Event event) {}
 
     protected void preUpdate() {}
 
@@ -61,13 +19,15 @@ public abstract class GameSystem
 
     protected void postUpdate() {}
 
-    public boolean isEnabled()
+    public final void update(float dt)
     {
-        return enabled;
-    }
+        preUpdate();
 
-    public void setEnabled(boolean value)
-    {
-        enabled = value;
+        for(Entity e : getEntities())
+        {
+            updateEntity(e, dt);
+        }
+
+        postUpdate();
     }
 }
